benerin front end

+++ View/kegiatanView.php
@@ -48,14 +48,28 @@ if (!isset($_SESSION['id_login'])){
 
 // Fetch all aktivitas data from database
 
-$result = mysqli_query($mysqli, "
+
+if ($_SESSION['role'] == "O"){
+
+    $result = mysqli_query($mysqli, "
     SELECT a.*, 
     (SELECT COUNT(*) FROM GABUNG_RELAWAN gr WHERE gr.A_ID = a.A_ID AND gr.GR_STATUS = 'Diterima') AS terdaftar
     FROM AKTIVITAS a WHERE O_ID = '".$_SESSION['o_id']."'");
+}
+
+else if ($_SESSION['role'] == "R"){
+
+    $result = mysqli_query($mysqli, "
+    SELECT a.*, 
+    (SELECT COUNT(*) FROM GABUNG_RELAWAN gr WHERE gr.A_ID = a.A_ID AND gr.GR_STATUS = 'Diterima') AS terdaftar
+    FROM AKTIVITAS a 
+    "
+);
+
+}
+
 
 ?>
-
-
 
 <body class="light sidebar-collapse sidebar-offCanvas-lg">
 <!-- Pre loader -->
@@ -66,7 +80,16 @@ $result = mysqli_query($mysqli, "
         <div class="navbar navbar-expand navbar-dark d-flex justify-content-between bd-navbar">
             <div class="relative offset-9" style="height: 50px;">
                 <form action="../Controller/authController.php" method="post">
-                    <span class="text-white">hello <b><?= $_SESSION['o_nama']; ?></b>!</span>
+                    <span class="text-white">hello 
+                    <?php 
+                    if($_SESSION['role'] == "O"){
+                        echo $_SESSION['o_nama']; 
+                    }
+                    else if($_SESSION['role'] == "R"){
+                        echo $_SESSION['r_nama']; 
+                    }
+                    ?>
+                    !</span>
                     <input type="hidden" name="act" value="logout">
                     <input type="submit" class="btn btn-danger" value="Logout">
                 </form>
@@ -87,14 +110,17 @@ $result = mysqli_query($mysqli, "
           <?php while($row = mysqli_fetch_object($result)){?>
           <div class="card my-3 no-b">
               <div class="card-body">
+              <?php if($_SESSION['role']=="O") { ?>
                   <div class="col-md-2 offset-11">
-                      <a class="btn btn-warning btn-sm" href="../View/editKegiatanView.php?a_id=<?= $row->A_ID ?>">
-                          <i class="icon-pencil"></i>
-                      </a>
-                      <a class="btn btn-danger btn-sm" href="../Controller/kegiatanController.php?a_id=<?= $row->A_ID ?>&act=delete">
-                          <i class="icon-trash"></i>
-                      </a>
+                  <a class="btn btn-warning btn-sm" href="../View/editKegiatanView.php?a_id=<?= $row->A_ID ?>">
+                  <i class="icon-pencil"></i>
+                  </a>
+                  <a class="btn btn-danger btn-sm" href="../Controller/kegiatanController.php?a_id=<?= $row->A_ID ?>&act=delete">
+                  <i class="icon-trash"></i>
+                  </a>
                   </div>
+               <?php  }?>  
+               
                   <div class="row d-md-flex align-items-center">
                       <div class="col-md-9 d-flex align-items-center">
                           <img class="mr-3 r-3" width="100" src="../assets/img/kegiatan/<?= $row->A_PATH ?>" alt="Gambar kegiatan" style="width: 400px; box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);">
@@ -142,12 +168,16 @@ $result = mysqli_query($mysqli, "
           <?php }?>
           <div class="card my-3 no-b">
               <div class="card-body">
+                <?php if ($_SESSION['role']=="O") {?> 
                   <div class="col-sm-12">
-                      <center><a type="button" class="btn btn-success btn-sm" href="tambahKegiatanView.php">
+                      <center>
+                      <a type="button" class="btn btn-success btn-sm" href="tambahKegiatanView.php">
                           <i class="icon-add"></i>
                           Tambah Kegiatan
-                      </a></center>
+                      </a>
+                      </center>
                   </div>
+                <?php  }?>
               </div>
           </div>
       </div>
